chore: restructure theme & styling
