+++ src/styles/constants/spacing.ts
@@ -1,0 +1,11 @@
+// ----------------------------------------------------------------------
+
+export const Spacing = {
+  half: 2,
+  one: 4,
+  two: 8,
+  three: 16,
+  four: 24,
+  five: 32,
+  six: 64,
+} as const;
