+++ src/styles/index.ts
@@ -1,0 +1,3 @@
+export * from './constants/colors';
+export * from './constants/fonts';
+export * from './constants/spacing';
